move ScriptExecutor
